+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGED:
 
 - moved jobject-core to its own project
 - update of test-objects dependency version to 5.1
+- new method for copy properties over reflection
 
 Version 3
 -------------